\name{GreedySBTM-package}
\alias{GreedySBTM-package}
\alias{GreedySBTM}
\docType{package}
\title{
\packageTitle{GreedySBTM}
}
\description{
\packageDescription{GreedySBTM}
}
\author{
Riccardo Rastelli

Mantainer: Riccardo Rastelli <riccardo.rastelli@wu.ac.at>
}
\references{
Rastelli, R. (2017) "Exact integrated completed likelihood maximisation in a stochastic block transition model for dynamic networks", \url{https://arxiv.org/abs/1710.03551}
}
\keyword{ package }

